% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChartChange14DaysIncidence.R
\name{Plot_Change14DaysIncidence}
\alias{Plot_Change14DaysIncidence}
\title{Plot Change 14Days Incidence}
\usage{
Plot_Change14DaysIncidence(Dataset, PopulationDataset, ListCountries)
}
\description{
This returns a plot displaying the change in 14days incidence for selected countries
}
\examples{
Plot_Change14DaysIncidence()
}
